Move HostConfiguration into Redwood
It's not wired up at the Redwood layer, but that's next.
