--- redwood-treehouse-configuration/build.gradle
@@ -1,22 +1,0 @@
-import app.cash.redwood.buildsupport.KmpTargets
-
-apply plugin: 'org.jetbrains.kotlin.multiplatform'
-apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
-apply plugin: 'app.cash.redwood.build'
-
-redwoodBuild {
-  publishing()
-}
-
-kotlin {
-  KmpTargets.addAllTargets(project)
-
-  sourceSets {
-    commonMain {
-      dependencies {
-        api projects.redwoodRuntime
-        api libs.kotlinx.serialization.core
-      }
-    }
-  }
-}
