+++ redwood-treehouse-host/src/androidUnitTest/kotlin/app/cash/redwood/treehouse/TreehouseWidgetViewTest.kt
@@ -27,6 +27,7 @@ import androidx.core.view.WindowInsetsCompat
 import app.cash.redwood.LayoutModifier
 import app.cash.redwood.treehouse.TreehouseView.WidgetSystem
 import app.cash.redwood.ui.Density
+import app.cash.redwood.ui.HostConfiguration
 import app.cash.redwood.ui.Margin
 import app.cash.redwood.widget.ViewGroupChildren
 import app.cash.redwood.widget.Widget
