+++ redwood-treehouse/src/commonMain/kotlin/app/cash/redwood/treehouse/ZiplineTreehouseUi.kt
@@ -16,6 +16,7 @@
 package app.cash.redwood.treehouse
 
 import app.cash.redwood.protocol.EventSink
+import app.cash.redwood.ui.HostConfiguration
 import app.cash.zipline.ZiplineService
 import kotlin.native.ObjCName
 import kotlinx.coroutines.flow.StateFlow
