+++ settings.gradle
@@ -72,7 +72,6 @@ include ':redwood-tooling-schema'
 include ':redwood-treehouse'
 include ':redwood-treehouse-composeui'
 include ':redwood-treehouse-composeui-insets'
-include ':redwood-treehouse-configuration'
 include ':redwood-treehouse-guest'
 include ':redwood-treehouse-guest-compose'
 include ':redwood-treehouse-host'
